Moving controllers to another project called Presentation to make architecture better (now we will not use Infrastructure in controllers) unless we needed to reference to all layers in the API layer for DependencyInjection
