+++ Sadin.Cms.sln
@@ -25,6 +25,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UnitTests", "UnitTests", "{
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sadin.Cms.Domain.UnitTests", "cms\tests\UnitTests\Sadin.Cms.Domain.UnitTests\Sadin.Cms.Domain.UnitTests.csproj", "{4FD7C554-AEB6-488B-B483-D93CEF91A42B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Website.Presentation", "cms\src\Website.Presentation\Website.Presentation.csproj", "{9137D8C6-D0CC-4B69-A0EA-209DCDEFCF55}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -62,6 +64,10 @@ Global
 		{4FD7C554-AEB6-488B-B483-D93CEF91A42B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4FD7C554-AEB6-488B-B483-D93CEF91A42B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4FD7C554-AEB6-488B-B483-D93CEF91A42B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9137D8C6-D0CC-4B69-A0EA-209DCDEFCF55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9137D8C6-D0CC-4B69-A0EA-209DCDEFCF55}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9137D8C6-D0CC-4B69-A0EA-209DCDEFCF55}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9137D8C6-D0CC-4B69-A0EA-209DCDEFCF55}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{11D5CB98-A4C2-45E2-99F7-A6DB7E6C6909} = {DB2BF6A9-A501-4F6A-AF05-06E7A5B74B5F}
@@ -74,5 +80,6 @@ Global
 		{0555D29A-39A6-491B-959D-ECB32E41A19F} = {DB2BF6A9-A501-4F6A-AF05-06E7A5B74B5F}
 		{7D0A9BF9-3DFC-4CD7-8EFF-B3A890E75BEC} = {0555D29A-39A6-491B-959D-ECB32E41A19F}
 		{4FD7C554-AEB6-488B-B483-D93CEF91A42B} = {7D0A9BF9-3DFC-4CD7-8EFF-B3A890E75BEC}
+		{9137D8C6-D0CC-4B69-A0EA-209DCDEFCF55} = {11D5CB98-A4C2-45E2-99F7-A6DB7E6C6909}
 	EndGlobalSection
 EndGlobal
