+++ cms/src/Website.Presentation/Controllers/SecureBaseController.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Website.Common.WebFrameworks;
 using Website.Services.AuthServices;
 
-namespace Website.Api.Controllers;
+namespace Website.Presentation.Controllers;
 
 [Authorize]
 public class SecureBaseController : BaseController
