+++ cms/src/Website.Presentation/ViewModels/ContactUs/ContactUsListItemViewModel.cs
@@ -1,28 +1,21 @@
-using Website.Application.ContactUsMessages.Queries.GetMessageById;
+using Website.Application.ContactUsMessages.Queries.GetPagedMessages;
 
-namespace Website.Api.ViewModels.ContactUs;
+namespace Website.Presentation.ViewModels.ContactUs;
 
-public class ContactUsMessageViewModel
+public class ContactUsListItemViewModel
 {
-    public ContactUsMessageViewModel(GetMessageByIdDto message)
+    public ContactUsListItemViewModel(GetPagedMessagesDto message)
     {
         Id = message.Id;
         FullName = message.FullName;
-        Email = message.Email;
-        PhoneNumber = message.PhoneNumber;
         Title = message.Title;
-        Content = message.Content;
         CreatedDate = message.CreatedDate;
         CheckedDate = message.CheckedDate;
         IsChecked = message.IsChecked;
     }
-    
     public Guid Id { get; private set; }
     public string FullName { get; private set; }
-    public string Email { get; private set; }
-    public string PhoneNumber { get; private set; }
     public string Title { get; private set; }
-    public string Content { get; private set; }
     public DateTimeOffset CreatedDate { get; private set; }
     public DateTimeOffset? CheckedDate { get; private set; }
     public bool IsChecked { get; private set; }
