+++ cms/src/Website.Presentation/Website.Presentation.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Website.Application\Website.Application.csproj" />
+    <ProjectReference Include="..\Website.Common\Website.Common.csproj" />
+    <ProjectReference Include="..\Website.Services\Website.Services.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="KSFramework" Version="1.1.24" />
+  </ItemGroup>
+
+</Project>
